@@ -1,37 +1,49 @@
 {
-  "name": "example-component",
-  "version": "1.0.0",
-  "description": "An example component with implementations of OIH actions",
-  "keywords": [
-    "elasticio-component"
-  ],
+  "homepage": "https://elastic.io",
   "author": {
-    "name": "Elastic.io GmbH",
-    "email": "ps@elastic.io",
-    "url": "https://www.elastic.io/"
+    "name": "elastic.io GmbH",
+    "email": "info@elastic.io",
+    "url": "https://www.elastic.io"
   },
-  "license": "Apache-2.0",
+  "keywords": [
+    "elasticio",
+    "elasticio-component",
+    "ipaas",
+    "integration",
+    "api"
+  ],
+  "engines": {
+    "node": "16.x"
+  },
   "scripts": {
-    "test": "mocha spec --recursive --timeout 50000",
-    "integration-test": "mocha spec-integration --recursive --timeout 50000"
+    "lint": "eslint --ext .ts --quiet --fix",
+    "pretest": "eslint --ext .ts --quiet --fix && find src spec spec-integration -name \"*.js\" -type f -delete && rm -f verifyCredentials.js",
+    "test": "nyc mocha --require ts-node/register \"spec/**/*spec.ts\" --timeout 10000",
+    "posttest": "tsc",
+    "integration-test": "npm run pretest && nyc mocha --require ts-node/register \"spec-integration/**/*spec.ts\" --timeout 500000"
   },
   "dependencies": {
-    "@elastic.io/component-commons-library": "1.1.5",
+    "axios": "0.27.2",
     "elasticio-node": "0.0.9",
-    "elasticio-sailor-nodejs": "2.6.18"
+    "elasticio-rest-node": "1.2.7",
+    "elasticio-sailor-nodejs": "2.6.27"
   },
   "devDependencies": {
     "@elastic.io/component-logger": "0.0.1",
-    "chai": "4.2.0",
+    "@types/mocha": "9.0.0",
+    "@types/node": "16.9.6",
+    "@typescript-eslint/eslint-plugin": "4.30.0",
+    "@typescript-eslint/parser": "4.30.0",
+    "chai": "4.3.4",
     "chai-as-promised": "7.1.1",
-    "dotenv": "8.2.0",
-    "eslint": "7.13.0",
+    "dotenv": "10.0.0",
+    "eslint": "7.32.0",
     "eslint-config-airbnb-base": "14.2.1",
-    "eslint-plugin-import": "2.22.1",
-    "mocha": "8.2.1",
-    "sinon": "9.2.1"
-  },
-  "engines": {
-    "node": "14"
+    "eslint-plugin-import": "2.24.2",
+    "mocha": "10.0.0",
+    "nyc": "15.1.0",
+    "sinon": "11.1.2",
+    "ts-node": "10.2.1",
+    "typescript": "4.4.3"
   }
 }
